backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[sqlite]
pytest
pyyaml